disconnected components from store
