+++ app/containers/Base/BaseContainer.js
@@ -1,21 +1,18 @@
 import React, { PropTypes } from 'react';
-import { lifecycle } from 'recompose';
 import { connect } from 'react-redux';
+import { lifecycle } from 'recompose';
+
+import ConnectContainer from '../../containers/ConnectContainer';
 
 import MainSidebar from '../../components/Main/Sidebar/MainSidebar';
-import ConnectSidebar from '../../components/Connect/Sidebar/ConnectSidebar';
-import ReduxFormBase from '../../components/Connect/Content/ReduxFormBase';
 import MainContent from '../../components/Main/Content/MainContent';
 import Header from '../../components/Base/Header/Header';
 import ContextMenu from '../../components/Base/ContextMenu/ContextMenu';
-import UpdatesModal from '../../components/Connect/Content/UpdatesModal/UpdatesModal';
 import { startAppMixpanelEvent } from '../../helpers';
-
 
 const propTypes = {
   isConnected: PropTypes.bool
 };
-
 
 const enhance = lifecycle({
   componentDidMount() {
@@ -23,32 +20,22 @@ const enhance = lifecycle({
   }
 });
 
-
 const BaseContainer = enhance(({ isConnected }) => (
   <div id="wrapper">
     <ContextMenu />
 
-    {isConnected ?
-      <div className="flex-row max-height">
+    {isConnected
+      ? <div className="flex-row max-height">
         <div className="flex-col left-pane">
           <Header />
           <MainSidebar />
         </div>
         <MainContent />
-      </div> :
-
-      <div className="flex-col max-height">
-        <Header />
-        <div className="flex-row max-height">
-          <UpdatesModal />
-          <ConnectSidebar />
-          <ReduxFormBase />
-        </div>
       </div>
+      : <ConnectContainer />
     }
   </div>
 ));
-
 
 BaseContainer.propTypes = propTypes;
 
@@ -57,5 +44,4 @@ function mapStateToProps(state) {
     isConnected: state.currentTable.isConnected
   };
 }
-
 export default connect(mapStateToProps, null)(BaseContainer);
